requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asttokens==2.4.1
    # via stack-data
colour==0.1.5
    # via varnaapi
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.1.1
    # via ipython
deprecated==1.2.14
    # via varnaapi
executing==2.0.1
    # via stack-data
fonttools==4.53.1
    # via matplotlib
ipython==8.26.0
    # via varnaapi
jedi==0.19.1
    # via ipython
kiwisolver==1.4.5
    # via matplotlib
lxml==5.2.2
    # via svgutils
matplotlib==3.8.4
    # via
    #   -r requirements.in
    #   seaborn
matplotlib-inline==0.1.7
    # via ipython
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   scipy
    #   seaborn
packaging==24.1
    # via matplotlib
pandas==2.2.2
    # via
    #   -r requirements.in
    #   seaborn
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.4.0
    # via
    #   -r requirements.in
    #   matplotlib
platformdirs==4.2.2
    # via varnaapi
prompt-toolkit==3.0.47
    # via ipython
psutil==6.0.0
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via ipython
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via varnaapi
scipy==1.13.0
    # via -r requirements.in
seaborn==0.13.2
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
stack-data==0.6.3
    # via ipython
svgutils==0.3.4
    # via -r requirements.in
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
tzdata==2024.1
    # via pandas
varnaapi==1.1.1
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
wrapt==1.16.0
    # via deprecated
